--- android/app/src/main/java/com/mulberry/ody/domain/repository/ody/FCMTokenRepository.kt
@@ -1,7 +1,0 @@
-package com.mulberry.ody.domain.repository.ody
-
-interface FCMTokenRepository {
-    suspend fun fetchFCMToken(): Result<String>
-
-    suspend fun postFCMToken(fcmToken: String)
-}
